requirements.txt
aniso8601==9.0.1
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.4
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-camel-case==1.3.0
drf-yasg==1.21.4
exceptiongroup==1.1.0
Faker==12.0.1
graphene==3.2.1
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.2
mixer==7.2.2
packaging==23.0
Pillow==9.3.0
pluggy==1.0.0
promise==2.3
pytest==7.2.1
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2022.7
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.14
